kensurbot: update trigger message with alive status if bot restarts
before restarting, the chat ID and message ID is stored in a file.
when the bot starts, it'll check if that file exists, and retrieve the
contents if it does. it'll be used to edit the "restarting" message.

this is much better than sending a message to botlog everytime the bot
starts.

Change-Id: I60b780caffdb02b3f748863c5e22efaff48f55e8
Signed-off-by: KenHV <yo@kenharris.xyz>