userbot: Add a workaround for Heroku's dyno migration
Heroku migrates the dyno randomly, and tries to be seamless about it.
This leads to the two instances overlapping, and Telegram invalidates
the session (AuthKeyDuplicatedError).

We'll attempt to work around this by storing the IP address on each
instance and checking if it's changed. If it is, we'll wait for 3
minutes before starting Telethon.

The timeout might not be enough, in which case it may be increased.

Change-Id: Ibed1581cfe2606f4a884ad9ab308b34d0f23a73f
Signed-off-by: KenHV <yo@kenharris.xyz>